move medications api to api router folder
